pyproject.toml
[project]
name = "manage-breast-screening-gateway"
version = "0.1.0"
description = "NHS Manage Breast Screening Gateway Service"
requires-python = ">=3.14,<4.0.0"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
  "pydicom>=3.0.1",
  "pynetdicom>=3.0.4",
  "numpy>=2.4.0,<3",
  "pylibjpeg>=2.0.0",
  "pylibjpeg-openjpeg>=2.5.0",
  "pillow>=11.0.0",
  "websockets==16.0",
  "requests>=2.32.0",
  "python-dotenv>=1.2.1",
  "azure-monitor-opentelemetry>=1.8.6",
]

[dependency-groups]
dev = [
  "pytest>=8.4.1,<10",
  "pytest-cov>=7.0.0,<8",
  "ruff>=0.14.1,<0.15",
  "pyright>=1.1.390",
  "ipdb>=0.13.13,<0.14",
  "pytest-asyncio>=1.3.0,<1.4.0",
  "python-semantic-release>=10.5.3",
]

[tool.uv]
required-version = ">=0.9.7"
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.semantic_release]
version_variable = ["pyproject.toml:version"]
upload_to_release = true
build_command = "true" # We use our own build script in Stage 3

[tool.semantic_release.branches.main]
match = "main"


[tool.semantic_release.changelog]
template_dir = "templates"
exclude_commit_patterns = [
  "chore\\(release\\):",
  "chore\\(deps-dev\\):",
  "chore\\(deps\\):",
]

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "github-actions <email@address.com>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov --cov-report=term-missing --cov-report=html --cov-report=xml"
markers = ["unit: Unit tests", "integration: Integration tests"]

[tool.ruff]
target-version = "py314"
line-length = 120
exclude = ["migrations"]
preview = true

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"] # Core + isort

[tool.ruff.lint.isort]
known-first-party = ["manage_breast_screening_gateway"]

[tool.coverage.run]
omit = ["*/tests/*", "**/test_*.py", "**/__main__.py"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if __name__ == .__main__.:",
  "raise AssertionError",
  "raise NotImplementedError",
]

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__", ".venv"]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
